lit-llm/pyproject.toml
[project]
name = "lit-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.72.6.post1",
    "python-dotenv>=1.1.0",
]
